First experiment about resizing popups at specific speeds "done"?
In this experiment, I implemented a mechanism where the speed at which the user resizes a popup is calculated. There are 5 levels, each with their own min speed and max speed. The player has to first resize very slowly, then level 2 faster and so forth. When the user is resizing too slowly, faster is written, and when too fast, slower is written. There is also a penalty that is subtracted to the number of growing points that add up when the user is resizing at the correct speed. This penalty is bigger as the level increases. There is also a continuous pleasure decrease of 1 point per frame. I implemented this because when the user is not resizing anymore, the pleasure slowly goes back down, like a normal person would basically. When you get turned on by someone touching you, you can continue and get more turned on if they do it right and more gradually more intensely, but also stop being turned on if they are really too intense fast (penalty goes really fast when the user resizes too fast) or slowly go down if the user is taking their time. The user starts with a pleasure gauge at 0 and it fills itself up to 100k, which closes the pop up and the player wins. The gauge is just a rectangle that grows in width from 0 to 1000 (mapped from 0 to 100k). Each of the levels have a target that the user has to reach in order to switch to the next set of rules (min and max speed, bigger penalties...) BUT there is a buffer where there are no penalties and automatic pleasure decrease for the player to adjust to the new speed of the new level. Once the player reaches the last target, the window closes and its the end.

I think the mechanics sort of work for now, but like i said in my journal, I feel like this is a) not funny and b) sort of exactly like "It is as if you were making love" by my supervisor Pippin Barr. Also there is no physics. The 5 levels for now don't make much sense, I just wanted to see if that could work in a context of levels.

So yeah, I'm not really satisfied by this thing and not too sure were I could bring it at this point. I feel like I have to brainstorm again about everything more generally conceptually.