created a veiled penis!
I created a veiled penis. The penis is a chain with many particles, so it feels sort of organic, like a tube of sorts. The balls are again two chains of two particles, with a static first one. I also added a thick line the color of the balls to create a sort of ballsack. I then added a constraint between both balls to keep them together as they would in a sack!! I felt happy with it, and had the idea of adding a cloth on top, to hide it. Not too sure why but I think there might be something funny and erotic about discovering it under the cloth?? So yeah not too sure where I'm heading with this if i'm honest haha.