Started vulva, started boobs, continued penis, updated journal
VULVA {
Added a clitoris-like particle with a static constraint, gives a bit of a feeling of a joystick ish. My first idea was to be able to play with it in order to make something else happen. The whole idea started because when I was working on my grieving game "Grief Rush Millenium" I played with making a matterjs chain with really tight constraints between particles and this produced a sort of unhinged vibrating creature moving around. I thought maaaybe I could use this as a vibrator of sorts. So I created a vibrator with two particles tightly contrained, and I though maybe at first the constraints are tight but not tight enough that they make the composite vibrate, but when the user picks the toy up, then the constraints tighten and it starts vibrating. The player then rubs it on the particle and something happens. I started doing that but wasn't really feeling anything. More on that in the journal. So I skipped to another experiment for the moment.
}

BOOBS {
Added two ellipses at first as boobs but I thought they were too round and not interesting. I was using two chain of boobs tightly constrained together with a line the same size drawn from particle to particle, with the first one being static, so that if I pull on them they seems to elongate a bit. I used the same principle as the penis but with tighter and shorter length of constraints.  I tried playing with new shapes, like a chain of polygons to try and mimick a more triangular shape. What I ended up doing was to make the particles gradually bigger in the chain. So the first static one was like 155px, then the next 5 px bigger and so forth, making a sort of pear-ish form which stayed simple and minimalistic like the other sketches but was still reminiscent of more realistic boobs (KINDOF). I also added nipples, with rectangular bodies with rounded corners (chamfer, new option I tried which I liked and got me excited ) At this point I was happy with the way they looked but I wasn’t too sure where I was headed (the game- toylike mechanics). Speaking with friends and showing it to them we started talking about electrostimulation (electrosex) and I thought this would be great. So I started working on making clamps (first time working with custom composites, like placing the rectangles at different positions and they all stick together as one thing (i should probably do this with the boobs and nipples when cleaning everything out just for consistency) and that was again super satisfying and fun and got me excited (making new things work). BUT the clamps were not really working in a way that I thought was interesting (they were hard to position and made me feel like the mechanic would be too complex compared to what I have going on in the other ones). So even though I liked what I had made technically and I’m probably going to use the concept somewhere elese, I started thinking about another option. I actually took a step back and was like : how am i even going to be able to stick these to the nipple (whatever “these” are). So i just created a simple circle particle and started searching for ways to “stick” it to another body. The thing is I wanted to be able to stick it off too. I google searched and found someone on a blog mentionning different ways they would go about doing that. So I started by calculating the distance first, and then tried with just setPosition, which did not work. I also played with applyForce towards the other body which did not work. Then I tried making it static, which also did not work (was just stuck on the position on the canvas even though the boob moved). But then I tried just setting the velocity to 0 so it stops moving. This should not have worked basically but because I had let the setPosition still there, IT worked. Like when the circle is close enough, bring the velocity to 0 and set its position to the other body. It stuck there. Now if I moved the boob too fast it would unstick to it, but I managed to fix that by making the dist bigger, so that it needs to move alot to fall off. Also its its nice because you have to pull on it a lot of make it unstick. I added a chain to this circle and a second circle at the other end and not I have a chain where each end is stuck to each boob which I think really rocks. Next thing: work on adding chaos/electricity when both ends are stuck. Cool. I have a good feeling about this one.
}

PENIS {

I’m following this person Instagram where they are doing the entire book “The Nature of Code” using p5 and showing all their small experiments. Weirdly they are working with matterjs atm, and they started doing a sort of liquid simulation using particles. I was like : how do you make that using circles and i need to make that because this is literally sperm. I made some research and found someone making exactly that but as milk coming out of a carton…….. So I spent a couple day trying to actually make this, but it was really cumbersome. After one day it wasn’t working and I was depressed and sort of put it to the side. But then I started again, and figured out how by looking at another similar example and realized I had to use css filters and not only JavaScript. I didn’t know exactly where I wanted the liquid to be but I new of course that the penis could be a hilarious option. So I did it there, and after a couple of weird things (playing with css, clearing the background of p5, putting a black background in css…) I finally got it. Well almost. The liquid was there, but it was affecting the entire canvas making the dick also gooey (which was also interesting but not what I wanted). I decided to make two p5 instances and have everything I had (matter js, drawing ….) on the background canvas, but drawing the particles (created on the background one) on the top canvas so I can put the SVG filter there. Now it was working but I couldnt play witht he dick anymore with the mouse because there was a whole canvas blocking the one underneath…….) SO on the third day, waking up, I thought ok im going to draw the penis on the background canvas BUT have the matterjs engine and bodies on the first one to have access to it. I started doing that but they REALIZED I could actually just change the mouse constraint to play with the bottom canvas when creating it so i literally just switched that and everything was working. it was so cool ngl. I’m kind of obsessed with it ahha. I removed the gland because it sort of fucked with the perspective of where the sperm goes out when lifting the dick. Now it gives almost a sort of 3D effect. I also adjusted the sperm origin position to follow the mouse up and down and left and right (slightly) so that the origin doesnt stay always in the center if you pull it to the sides or top or bottom, making it even more realistic. Another thing I added that is really cool, is the force of the sperm particles are also following the mouse. Before they were just falling down whichever way you were pulling the dick. Now when you pull it to the right, it shoots it to the right and so forth in every direction. I maybe want to add that if you pull stronger the force is also more intense but we shall see. Last thing I added is the wrapping effect using a plugin (I had no idea these existed). I added a link through a script, loaded the plugin and just used their function to make all the particles wrap around the sides. Its impossible to have only the particles wrap from the sides (needs to be top and bottom as well), but I adjusted the y max (the point where they wrap back up top) to be bigger that the point where they are actually removed from the world and the array, so they only loop from the sides. I also added a second stream of smaller particles to make it more liquidesque. Im obsessed by the realism/playfulness/grotesque/ exageration of it all. I think its really confusing. Its hilarious. I live.

}