Added a slapping booty when clicking on cheek, some sounds to anal beads sketch and new colors!!!!
Slapping booty: I wanted to add one more fun interaction and I thought this could be nice. I used a matter event to detect when clicking on one of the cheeks, add applied force to the clicked on cheek to make it bounce, mimicking a big bouncing booty.

Sounds: The only two things I was sure was the pop sound when getting a bead out and a slap sound when clicking on the bum. I also maybe wanted a sound when pulling on the beads (have not tested that yet but I think this is not going to work well). I went to splice.com and download many bodily sounds and ended up choosing some really simple cartoony sounds. The popping sound was perfect, I love the effect, and it seems to even intensify the sort of sucction aspect. I didnt like a specific slap sounds so i booted Ableton and tried to stack and EQ slaps sounds to make the perfect one, but the sound started to feel really complex compared to the pop sound that was so simple and minimalist, and with the visuals as well. Every time I try to over process or add complexity to these experiments, it falls short and becomes less funny I feel like. I think a lot of the humour comes from the mix between the explicit sexual nature of the interaction + the minimalism of the visuals, color choices and sound (which reminds of cartoonish or not sexual things usually). Its also very design-y, meaning that it is cold, like an illustration. There is no realism in shapes or shading. It is just cold almost brutalist minimalist design which kind of takes out part of the erotic experience and gives more place to humour.

Visuals: I kinda new from the beginning that I wanted a very colorful palette with thin strokes around some shapes. I'm really inspired by @its_aless_mc on Instagram (mtl based designer) for the color and the naive style. I think its really coming together with visuals and sounds.