Added the good text of club sexu, switched back from shrinking the beads when clicking to always big as long as they are being hovered, ALSO big change but changed the info beads wich were actualy just javascript p5 and the text was an image to divs. There is also 10 beads now to fit all the information Club Sexu gave me. And I added an end.
So i was talking with club sexu and we decided to use the beads as a sort of platul interface to interact with parts of specific related articles to invite people to read the full texts. The anal beads one is to play with an article on sex toys, specifically anal beads. I tried a version with an image of placeholder text and it was nice. They also liked it. I'm a bit wary that it is taking away from the minimalist aesthetic which I think is where a lot of the humour comes from, everytime i add a visual thing i feel like im breaking the fluidity of the piece. I also liked that it was a random useless internet toy. But hey I think its actually pretty cool to make it something useful like a way to interact with the text. I think it opens up a lot of ideas and possibilites, altough I think now it falls more into the "clever" category and less the funny joke?? I dont know, maybe im to used to playing with it at this point. I think using divs instead of JS is nice because I can change the text easily and theres no problem with resolution and such. I think animations are smoother as well. I added beads to fit all the paragraphs given by club sexu. I also added an end by giving the link to the full article that appears when the whole toy is gone out of frame. I was contemplating making an infinite anal bead i thought that could be fun. But no things must be useful!!! than more people see them. I hope a lot of people see this and that it is as cool as I force myself to think that it is. Like viral cool. I think as a series its fun. Sexual playful interfaces to interact with text. I need to think about typography as wellll.