penis? Yes I worked on penis. Reflecting on how to play with text and use the penis as an interface to interact with portions of a club sexu article, implementing sperm containers (haha), reducing the size of the penis, putting the canvas in the middle of the window
So because the project is now something that is useful (an interace to interact with parts of an article), I started thinking about how I could do this using the penis and or the sperm. I really wanted to try to fill something up witht the liquid, before the sperm was only going everywhere and with the plugin wrap it goes back on the other side when leaving screen. Its nice but i think the idea of filling up a container is interesting. I pretty much fluidly (lol) had the idea of revealing the text by filling an invisible container. So the text is the color of the background in the middle of the invisible container, so you see nothing. But when filling up the container, it becomes white, and then the text appears because its on a white spermy backgound. I have 4 containers in each corner of the square canvas, which is nice because now you have to sort of shoot or fiigure out how to fill them to get all 4 facts. Im planning on putting a symbol like 3 small lines in each corner to signal that there is text there somehow, and i plan on doing so with the anal beads as well, creating a visual coherence between both of the interfaces. Wow. i really wonder what the third boob interface will become, feels like im going deep in those two now. We shall see.

I also reduced the size of the penis because it felt too confusing with the containers. now its easier to manipulate. I kinda miss the big dong though. I thought it was nice. Yeah. :'(

I put the canvas in the middle of the window, and changed the width and size and now I feel its really coherent with the anal beads one. Im really not sure about colors though. I will probably ask people who are good at that and try stuff.

I like this. Its so wrong. Glasses of sperm really. wow.