Penis interface? are you there?
I'm trying to add the text sent by Club Sexu from the article they want to showcase through this toy. I've put one paragraph in one of the 4 cases in each corner, which should reveal its text when being filled with sperm... (this feels wrong).
The text is definetly too long now so will have to adjust with them if they decide to answer me at some point. I succeeded in working with divs and text and aligning them with my containers, even when resizing the window. THE THING is that i'm working with two canvases on top of each other to have the liquid effect of the sperm (top canvas) and the normal shapes of the penis (bottom one). All my matter.js stuff is still created and manipulated on the bottom one but the top one is just displaying sperm with a SVG filter. I added the text divs to the second one but realisezed they would be hidden by sperm instead of revealed even with playing with z-index. I always hate introducing html and css in my JS work because I feel like i don't have a full grasp on those and it will fuck up the project but when working with text i think its best. ANYWAY now that im writing this i'm realising that I actually could append them to the second one and try to cancel the SVG on the divs? WIll try this later.

Also, im wondering if this is too gross? I think maybe being gay and gay culture is so gross that I think filling cups with sperm is fine? I mean its definetly subversive and surprising but is it funny? I think its funny in not a ahahahaha kind of way but more like HA... HA. OMG HA. and like thinking about it after and being like this was so funny and weird but without laughing? I wonder if club sexu will want tot share that on their page though.
Last thing is the performance aspect to this, because filling the sperm cups necessitates a lot of particles on the screen, its sort of slowing down the program when filling the last one. Dunno how to fix this. Also my computer is pretty old now but also still a beast. Will have to test on something else.

Also not so sure about the colors anymore on this. They annoye me.