Changed colors for penis sketch, introduced html info card for the penis sketch, made a small recipient to gather the sperm, adjusted the matter js options for more realism
I changed the colors for the ones my friend Julien gave me, which are awesome and so beautiful. I'm just not great at doing that, and he did an amazing job. I think I'm good with colors when i have complete freedom, but because i'm trying to make it fit with the branding of Club Sexu, I was not able to make something looking nice. The colors really add a layer of pro to the toys, and make them even more engaging and invite curiosity. I quite like them now, like a breath of fresh air.

I reused the HTML info card I previously had for the anal beads, I'm trying to keep some coherence design-wise. I positioned it at the bottom and moved the penis up, and also changed the canvas for the same size and proportions of the anal beads for consistency and also because if i ever try to make this mobile friendly then it will be less work.

I built a small recipient the size of the html info card to gather the sperm at the bottom of the canvas. I havent figuered out how I'm going to make the text appear with the sperm because now the background is gradient, so it can't be the same color as the background. I'll look into it tonight. I decided to go with one card at a time, and like the anal beads, there is a continue button that players can click, and that would make the bottom border disappear, emptying the sperm cup and then going to the next card, so the user fills again.

There was a lot going on matter js wise, like the gravity and mass of a lot of stuff was weird, so I decided to bring everything back to normal and try to adjust all the elements to get back to something realistic. I feared that the way it was working asked for a lot of performance. Also I think it couldve broken the program at some point.

I updated the journal about my whole journey about trying to figure out an SVG alternative because its not mobile friendly.